Various docs
- Updated permission Nullable
- migrating Integer to int in CustomStack
- fixed inconsistent HUD classes naming
- added PlayerQuantityHudWrapper docs
