+++ src/main/java/dev/lone/itemsadder/api/FontImages/PlayerHudsHolderWrapper.java
@@ -1,10 +1,14 @@
 package dev.lone.itemsadder.api.FontImages;
 
 import dev.lone.itemsadder.api.NotActuallyItemsAdderException;
+import org.bukkit.entity.Player;
 
-public class PlayerHudWapper
+/**
+ * Class representing a container which holds all HUDs of a player.
+ */
+public class PlayerHudsHolderWrapper
 {
-    public PlayerHudWapper(PlayerHUDsHolderWrapper playerHUDsHolderWrapper, String namespacedID)
+    public PlayerHudsHolderWrapper(Player player)
     {
         throw new NotActuallyItemsAdderException();
     }
@@ -17,17 +21,19 @@ public class PlayerHudWapper
         throw new NotActuallyItemsAdderException();
     }
 
-    public String getNamespacedID()
+    /**
+     * Get owner of this holder.
+     * @return Owner of this holder.
+     */
+    public Player getPlayer()
     {
         throw new NotActuallyItemsAdderException();
     }
 
-    public boolean exists()
-    {
-        throw new NotActuallyItemsAdderException();
-    }
-
-    public void setVisible(boolean value)
+    /**
+     * Force the recalculation the offsets if you shown a new HUD, added/removed new font images to a {@link PlayerCustomHudWrapper} HUD.
+     */
+    public void recalculateOffsets()
     {
         throw new NotActuallyItemsAdderException();
     }
