+++ src/main/java/dev/lone/itemsadder/api/FontImages/PlayerQuantityHudWrapper.java
@@ -1,11 +1,13 @@
 package dev.lone.itemsadder.api.FontImages;
 
-
 import dev.lone.itemsadder.api.NotActuallyItemsAdderException;
 
-public class PlayerQuantityHudWapper extends PlayerHudWapper
+/**
+ * Class representing a HUD which accepts float values and its textures are handled by ItemsAdder.
+ */
+public class PlayerQuantityHudWrapper extends PlayerHudWrapper
 {
-    public PlayerQuantityHudWapper(PlayerHUDsHolderWrapper playerHUDsHolderWrapper, String namespacedID)
+    public PlayerQuantityHudWrapper(PlayerHudsHolderWrapper playerHUDsHolderWrapper, String namespacedID)
     {
         super(playerHUDsHolderWrapper, namespacedID);
         throw new NotActuallyItemsAdderException();
@@ -19,11 +21,21 @@ public class PlayerQuantityHudWapper extends PlayerHudWapper
         throw new NotActuallyItemsAdderException();
     }
 
+    /**
+     * Set the float value of this HUD.
+     *
+     * @param value
+     *        Float value to set.
+     */
     public void setFloatValue(float value)
     {
         throw new NotActuallyItemsAdderException();
     }
 
+    /**
+     * Get the float value of this HUD.
+     * @return The float value of this hud.
+     */
     public float getFloatValue()
     {
         throw new NotActuallyItemsAdderException();
